Detailed log of credit usage #96

+++ database/migrations/2019_07_31_000014_create_bot_instances_table.php
@@ -18,10 +18,10 @@ class CreateBotInstancesTable extends Migration
             $table->unsignedInteger('user_id');
             $table->unsignedInteger('bot_id')->nullable();
             $table->unsignedInteger('aws_region_id')->nullable();
-            $table->double('used_credit')->default(0);
-            $table->double('up_time')->default(0);
-            $table->double('total_up_time')->default(0);
-            $table->double('cron_up_time')->default(0);
+            $table->unsignedInteger('used_credit')->default(0);
+            $table->unsignedInteger('up_time')->default(0);
+            $table->unsignedInteger('total_up_time')->default(0);
+            $table->unsignedInteger('cron_up_time')->default(0);
             $table->boolean('is_in_queue')->default(1);
             $table->enum('aws_status', [
                 'pending',
